refactor: rename build to make and move buildOptionsFromMessage function
