+++ src/cards/BaseCard.ts
@@ -9,30 +9,27 @@ interface Card {
 
   buildFooter?(): void;
 
-  build(): chatV1.Schema$CardWithId;
+  make(): chatV1.Schema$CardWithId;
 }
 
 export abstract class BaseCard implements Card {
   private id: string = 'cardId';
-  private _header: chatV1.Schema$CardHeader = {};
   private _content: chatV1.Schema$GoogleAppsCardV1Section[] = [];
-  private _footer: chatV1.Schema$GoogleAppsCardV1CardFixedFooter = {};
 
-  protected card = {
-    'header': this._header,
-    'sections': this._content,
-    'fixedFooter': this._footer,
+  protected card: chatV1.Schema$GoogleAppsCardV1Card = {
+    sections: this._content,
   };
 
   abstract buildSections(): void;
 
-  build(): chatV1.Schema$CardWithId {
+  make(): chatV1.Schema$CardWithId {
     return {
       'cardId': this.id,
       'card': this.card,
     };
   }
+
   buildMessage(): chatV1.Schema$Message {
-    return {cardsV2: [this.build()]};
+    return {cardsV2: [this.make()]};
   }
 }
